Refactor validator testing to validators_test package.
