+++ pkg/validate/validate_test.go
@@ -5,33 +5,8 @@ import (
 	"testing"
 
 	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestAllValidators(t *testing.T) {
-	for _, validator := range validators.TestRegistry.All() {
-		validator := validator
-		t.Run(validator.Name(), func(t *testing.T) {
-			t.Parallel()
-			// testing the succeeding candidates
-			succeedingMetaBundles := validator.SucceedingCandidates()
-			for _, mb := range succeedingMetaBundles {
-				res := validator.Run(mb)
-				require.False(t, res.IsError())
-				assert.True(t, res.IsSuccess())
-			}
-
-			// (optional) testing the failing candidates
-			failingMetaBundles := validator.FailingCandidates()
-			for _, mb := range failingMetaBundles {
-				res := validator.Run(mb)
-				require.False(t, res.IsError())
-				assert.False(t, res.IsSuccess())
-			}
-		})
-	}
-}
 
 func TestFilterDisabledValidators(t *testing.T) {
 	n_validators := validators.Registry.Len()
