+++ pkg/validators/am0001_test.go
@@ -1,9 +1,10 @@
-package validators
+package validators_test
 
 import (
 	"github.com/mt-sre/addon-metadata-operator/api/v1alpha1"
 	"github.com/mt-sre/addon-metadata-operator/internal/testutils"
 	"github.com/mt-sre/addon-metadata-operator/pkg/types"
+	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
 )
 
 func init() {
@@ -13,11 +14,11 @@ func init() {
 type TestAM0001 struct{}
 
 func (t TestAM0001) Name() string {
-	return AM0001.Name
+	return validators.AM0001.Name
 }
 
 func (t TestAM0001) Run(mb types.MetaBundle) types.ValidatorResult {
-	return AM0001.Runner(mb)
+	return validators.AM0001.Runner(mb)
 }
 
 func (t TestAM0001) SucceedingCandidates() []types.MetaBundle {
