+++ pkg/validators/am0002_test.go
@@ -1,8 +1,9 @@
-package validators
+package validators_test
 
 import (
 	"github.com/mt-sre/addon-metadata-operator/api/v1alpha1"
 	"github.com/mt-sre/addon-metadata-operator/pkg/types"
+	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
 )
 
 func init() {
@@ -12,11 +13,11 @@ func init() {
 type TestAM0002 struct{}
 
 func (val TestAM0002) Name() string {
-	return AM0002.Name
+	return validators.AM0002.Name
 }
 
 func (val TestAM0002) Run(mb types.MetaBundle) types.ValidatorResult {
-	return AM0002.Runner(mb)
+	return validators.AM0002.Runner(mb)
 }
 
 func (val TestAM0002) SucceedingCandidates() []types.MetaBundle {
