+++ pkg/validators/am0006_test.go
@@ -1,10 +1,11 @@
-package validators
+package validators_test
 
 import (
 	"github.com/mt-sre/addon-metadata-operator/api/v1alpha1"
 	"github.com/mt-sre/addon-metadata-operator/internal/testutils"
 	mtsrev1 "github.com/mt-sre/addon-metadata-operator/pkg/mtsre/v1"
 	"github.com/mt-sre/addon-metadata-operator/pkg/types"
+	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
 )
 
 func init() {
@@ -14,11 +15,11 @@ func init() {
 type TestAM0006 struct{}
 
 func (t TestAM0006) Name() string {
-	return AM0006.Name
+	return validators.AM0006.Name
 }
 
 func (t TestAM0006) Run(mb types.MetaBundle) types.ValidatorResult {
-	return AM0006.Runner(mb)
+	return validators.AM0006.Runner(mb)
 }
 
 func (t TestAM0006) SucceedingCandidates() []types.MetaBundle {
