+++ pkg/validators/am0008_test.go
@@ -1,9 +1,10 @@
-package validators
+package validators_test
 
 import (
 	"github.com/mt-sre/addon-metadata-operator/api/v1alpha1"
 	"github.com/mt-sre/addon-metadata-operator/internal/testutils"
 	"github.com/mt-sre/addon-metadata-operator/pkg/types"
+	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
 )
 
 func init() {
@@ -13,11 +14,11 @@ func init() {
 type TestAM0008 struct{}
 
 func (val TestAM0008) Name() string {
-	return AM0008.Name
+	return validators.AM0008.Name
 }
 
 func (val TestAM0008) Run(mb types.MetaBundle) types.ValidatorResult {
-	return AM0008.Runner(mb)
+	return validators.AM0008.Runner(mb)
 }
 
 func (val TestAM0008) SucceedingCandidates() []types.MetaBundle {
