+++ pkg/validators/registry.go
@@ -45,30 +45,3 @@ func (r defaultRegistry) Get(k string) (types.Validator, bool) {
 	v, ok := r.Data[k]
 	return v, ok
 }
-
-// TestRegistry - register all test structs
-var TestRegistry = NewTestRegistry()
-
-func NewTestRegistry() *testRegistry {
-	return &testRegistry{
-		&sync.Mutex{},
-		[]types.ValidatorTest{},
-	}
-}
-
-type testRegistry struct {
-	*sync.Mutex
-	Data []types.ValidatorTest
-}
-
-// Add - update the test registry in a thread-safe way. Called in init() functions
-func (t *testRegistry) Add(v types.ValidatorTest) {
-	t.Lock()
-	defer t.Unlock()
-
-	t.Data = append(t.Data, v)
-}
-
-func (t *testRegistry) All() []types.ValidatorTest {
-	return t.Data
-}
